用æfinish the user management Page and profile modify page

+++ src/main/java/com/fh/mall/entity/TestUser.java
@@ -3,14 +3,14 @@ package com.fh.mall.entity;
 import lombok.Data;
 
 /**
- * @Description: TODO(测试JqGrid插件的一个实体类，用完之后可以删除)
+ * @Description: 测试JqGrid插件的一个实体类，用完之后可以删除
  * @Author HueFu
  * @Date 2020-8-11 20:39
  */
 @Data
 public class TestUser {
 
-    private long id;
+    private int id;
     private String name;
     private String password;
